# ClinicService/Models/Requests/UpdatePetRequest.cs
namespace WebApplication10.Models.Requests
{
    public class UpdatePetRequest
    {
        public int PetId { get; set; }

        public int ClientId { get; set; }


        public string Name { get; set; }

        public DateTime Birthday { get; set; }
    }
}